pyproject.toml
[tool.poetry]
name = "python-poetry-docker-example"
version = "0.0.1"
description = ""
authors = ["Jota juandgomez13@gmail.com"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^3.0.3"
SQLAlchemy = "^2.0.29"
Werkzeug = "^3.0.2"
Jinja2 = "^3.1.3"
MarkupSafe = "^2.1.5"
mysqlclient = "^2.2.4"
Flask-SQLAlchemy = "^3.1.1"
flask-sqlacodegen = "^2.0.0"
mysql-connector-python = "^8.4.0"
Flask-WTF = "^1.2.1"
python-dotenv = "^1.0.1"
